Fix non-deterministic error in java email service test
This commit fixes an issue where the JavaEmailServiceTest sometimes fails due to a jakarta.mail.MessagingException. The issue is encountered in the local dev environment. This is fixed by explicitly setting the from field in the MimeMessage. However, and in production, the from field should not be set as it is authenticated. The field can only be set in testing and via reflection.